Add CI/CD and brought repository into compliance with modern standards
