+++ .bonsai/Setup.cmd
@@ -1,0 +1,4 @@
+@echo off
+pushd %~dp0
+powershell -ExecutionPolicy Bypass -File ./Setup.ps1
+popd
