+++ .editorconfig
@@ -1,0 +1,46 @@
+root = true
+
+[*]
+indent_style = space
+
+#--------------------------------------------------------------------------------------------------
+# XML, JSON, and web files
+#--------------------------------------------------------------------------------------------------
+[*.{xml,csproj,vcxproj,vcxproj.filters,shproj,props,targets,config,nuspec,resx,vsixmanifest,wxs,vstemplate,slnx}]
+indent_size = 2
+
+[*.json]
+indent_size = 2
+
+[*.{html,css}]
+indent_size = 2
+
+#--------------------------------------------------------------------------------------------------
+# C++
+#--------------------------------------------------------------------------------------------------
+[*.{c,cpp,h,hpp,ixx}]
+indent_size = 4
+charset = utf-8-bom
+trim_trailing_whitespace = true
+insert_final_newline = true
+
+#--------------------------------------------------------------------------------------------------
+# C#
+#--------------------------------------------------------------------------------------------------
+[*.{cs,csx}]
+indent_size = 4
+charset = utf-8-bom
+trim_trailing_whitespace = true
+insert_final_newline = true
+
+# Language keyword vs full type name
+# Predefined for members, etc does not create a message because the explicitly sized types are conveient in interop scenarios where the bit size matters.
+dotnet_style_predefined_type_for_locals_parameters_members = true:none
+dotnet_style_predefined_type_for_member_access = true:suggestion
+
+# Instantiate argument exceptions correctly
+dotnet_diagnostic.CA2208.severity = warning
+
+# Don't complain about not using modern collection syntax
+dotnet_style_prefer_collection_expression = never
+csharp_style_prefer_range_operator = false
