+++ .gitignore
@@ -1,3 +1,5 @@
-.vs
-bin
-obj
+.bonsai/Bonsai.exe*
+.bonsai/Packages/
+.bonsai/Settings/
+.vs/
+/artifacts/
