--- Bonsai.PointGrey/Bonsai.PointGrey.csproj
@@ -1,37 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Authors>Gonçalo Lopes</Authors>
-    <Title>Bonsai - PointGrey Library</Title>
-    <Description>Bonsai Library containing modules for acquiring images from PointGrey cameras.</Description>
-    <Copyright>Copyright © Gonçalo Lopes 2021</Copyright>
-    <ProjectUrl>https://github.com/bonsai-rx/pointgrey</ProjectUrl>
-    <PackageLicenseUrl>https://github.com/bonsai-rx/pointgrey/blob/master/LICENSE</PackageLicenseUrl>
-    <PackageIconUrl>https://bonsai-rx.org/assets/images/bonsai.png</PackageIconUrl>
-    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-    <PackageTags>Bonsai Rx PointGrey FlyCapture</PackageTags>
-    <IncludeSymbols Condition="'$(Configuration)'=='Release'">true</IncludeSymbols>
-    <GeneratePackageOnBuild Condition="'$(Configuration)'=='Release'">true</GeneratePackageOnBuild>
-    <PackageOutputPath>..\bin\$(Configuration)</PackageOutputPath>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <TargetFramework>net462</TargetFramework>
-    <Features>strict</Features>
-    <Version>2.5.0</Version>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Content Include="Bonsai.PointGrey.props" PackagePath="build\net462" />
-    <Content Include="Bonsai.PointGrey.targets" PackagePath="build\net462" />
-    <Content Include="..\Externals\FlyCapture2Managed\net\**\*.dll" PackagePath="build\net462\bin" />
-    <Content Include="..\Externals\FlyCapture2Managed\native\**\*.dll" PackagePath="build\native\bin" />
-    <Reference Include="FlyCapture2Managed_v100">
-      <HintPath>..\Externals\FlyCapture2Managed\net\x64\FlyCapture2Managed_v100.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Bonsai.Core" Version="2.5.0" />
-    <PackageReference Include="OpenCV.Net" Version="3.3.1" />
-  </ItemGroup>
-
-</Project>
