--- Bonsai.PointGrey/Properties/AssemblyInfo.cs
@@ -1,7 +1,0 @@
-﻿using Bonsai;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: XmlNamespacePrefix("clr-namespace:Bonsai.PointGrey", "pgr")]
-[assembly: WorkflowNamespaceIcon("Bonsai:ElementIcon.Video")]
