--- Bonsai.PointGrey/Properties/launchSettings.json
@@ -1,9 +1,0 @@
-{
-  "profiles": {
-    "Bonsai": {
-      "commandName": "Executable",
-      "executablePath": "$(registry:HKEY_CURRENT_USER\\Software\\Goncalo Lopes\\Bonsai@InstallDir)Bonsai.exe",
-      "commandLineArgs": "--lib:$(TargetDir)."
-    }
-  }
-}
