+++ Directory.Build.props
@@ -1,0 +1,3 @@
+﻿<Project>
+  <Import Project="build/Common$(MSBuildProjectExtension).props" Condition="Exists('build/Common$(MSBuildProjectExtension).props')"/>
+</Project>
