+++ Directory.Build.targets
@@ -1,0 +1,3 @@
+﻿<Project>
+  <Import Project="build/Common$(MSBuildProjectExtension).targets" Condition="Exists('build/Common$(MSBuildProjectExtension).targets')" />
+</Project>
