+++ docs/filter.yml
@@ -1,0 +1,4 @@
+apiRules:
+- exclude:
+    hasAttribute:
+      uid: System.ObsoleteAttribute
