+++ docs/template/public/main.css
@@ -1,0 +1,1 @@
+@import "workflow.css";
