+++ docs/template/public/main.js
@@ -1,0 +1,13 @@
+import WorkflowContainer from "./workflow.js"
+
+export default {
+    defaultTheme: 'light',
+    iconLinks: [{
+        icon: 'github',
+        href: 'https://github.com/bonsai-rx/pointgrey',
+        title: 'GitHub'
+    }],
+    start: () => {
+        WorkflowContainer.init();
+    }
+}
