+++ docs/toc.yml
@@ -1,0 +1,2 @@
+- name: API
+  href: ../artifacts/docs/api/
