+++ docs/workflows/.gitignore
@@ -1,0 +1,2 @@
+*.layout
+*.svg
