+++ src/Bonsai.PointGrey/Bonsai.PointGrey.csproj
@@ -1,0 +1,24 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <Title>Bonsai - PointGrey Library</Title>
+    <Description>Bonsai Library containing modules for acquiring images from PointGrey cameras.</Description>
+    <TargetFramework>net462</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Content Include="Bonsai.PointGrey.props" PackagePath="build\net462" />
+    <Content Include="Bonsai.PointGrey.targets" PackagePath="build\net462" />
+    <Content Include="..\Externals\FlyCapture2Managed\net\**\*.dll" PackagePath="build\net462\bin" />
+    <Content Include="..\Externals\FlyCapture2Managed\native\**\*.dll" PackagePath="build\native\bin" />
+    <Reference Include="FlyCapture2Managed_v100">
+      <HintPath>..\Externals\FlyCapture2Managed\net\x64\FlyCapture2Managed_v100.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Bonsai.Core" Version="2.5.0" />
+    <PackageReference Include="OpenCV.Net" Version="3.3.1" />
+  </ItemGroup>
+
+</Project>
