+++ src/Externals/FlyCapture2Managed/net/x86/FlyCapture2Managed_v100.dll
